+++ packages/contracts-communication/contracts/events/InterchainClientV1Events.sol
@@ -3,7 +3,7 @@ pragma solidity ^0.8.0;
 
 abstract contract InterchainClientV1Events {
     event ExecutionFeesSet(address executionFees);
-    event LinkedClientSet(uint256 chainId, bytes32 client);
+    event LinkedClientSet(uint64 chainId, bytes32 client);
 
     // TODO: figure out indexing
