+++ packages/contracts-communication/contracts/interfaces/IInterchainModule.sol
@@ -5,9 +5,9 @@ pragma solidity ^0.8.0;
 /// therefore this is not a part of a common interface.
 interface IInterchainModule {
     error InterchainModule__NotInterchainDB(address caller);
-    error InterchainModule__IncorrectSourceChainId(uint256 chainId);
+    error InterchainModule__IncorrectSourceChainId(uint64 chainId);
     error InterchainModule__InsufficientFee(uint256 actual, uint256 required);
-    error InterchainModule__SameChainId(uint256 chainId);
+    error InterchainModule__SameChainId(uint64 chainId);
 
     /// @notice Request the verification of a batch from the Interchain DataBase by the module.
     /// If the batch is not yet finalized, the verification on destination chain will be delayed until
