+++ packages/contracts-communication/test/apps/InterchainAppV1.t.sol
@@ -72,7 +72,7 @@ abstract contract InterchainAppV1Test is Test, AbstractICAppEvents, InterchainAp
         emit AppConfigV1Set(config.requiredResponses, config.optimisticPeriod);
     }
 
-    function expectEventAppLinked(uint256 chainId, bytes32 remoteApp) internal {
+    function expectEventAppLinked(uint64 chainId, bytes32 remoteApp) internal {
         vm.expectEmit(address(appHarness));
         emit AppLinked(chainId, remoteApp);
     }
@@ -142,11 +142,11 @@ abstract contract InterchainAppV1Test is Test, AbstractICAppEvents, InterchainAp
         vm.expectRevert(abi.encodeWithSelector(AbstractICApp.InterchainApp__NotInterchainClient.selector, account));
     }
 
-    function expectRevertReceiverNotSet(uint256 chainId) internal {
+    function expectRevertReceiverNotSet(uint64 chainId) internal {
         vm.expectRevert(abi.encodeWithSelector(AbstractICApp.InterchainApp__ReceiverNotSet.selector, chainId));
     }
 
-    function expectRevertSameChainId(uint256 chainId) internal {
+    function expectRevertSameChainId(uint64 chainId) internal {
         vm.expectRevert(abi.encodeWithSelector(AbstractICApp.InterchainApp__SameChainId.selector, chainId));
     }
